+++ .test/resourceIndex/index.ts
@@ -1,1 +1,2 @@
 export * from './MyClass';
+export {ExportedAsAlias as ExportAlias} from './SpecialExports';